Rewrite main checking function using promises to allow for checking more than one channel
Rewrite of main checking function using promises. This allows for the status of multiple channels to be checked.

Each channel name is loaded via a specified array. Each status check is loaded into a promise array and executed at the same time. Discord icon will be uploaded if at least one channel was found to be online, and set to offline if none are found.

Updated discord.js and tmi.js versions.
Added bluebird and lodash.

This implementation should overall be more reliable than the previously reverted commit.